DL: restructure dl for jax_backend
