+++ examples/dl/cnn_mnist.py
@@ -61,9 +61,10 @@ from sklearn.metrics import (
 import matplotlib.pyplot as plt
 from tensorflow.keras.datasets import mnist
 
-from scirex.core.dl import Model, Network, layers, activations, utils
-from scirex.core.dl.losses import cross_entropy_loss
-from scirex.core.dl.metrics import accuracy
+from scirex.core.dl.jax_backend.equinox import layers, activations, utils
+from scirex.core.dl.jax_backend.equinox.base import Model, Network
+from scirex.core.dl.jax_backend.equinox.losses import cross_entropy_loss
+from scirex.core.dl.jax_backend.equinox.metrics import accuracy
 
 # Set random seed for reproducibility
 key = jax.random.PRNGKey(42)
