+++ scirex/core/dl/jax_backend/equinox/__init__.py
@@ -1,0 +1,9 @@
+# from . import layers
+# from . import networks
+# from . import activations
+# from . import losses
+# from . import metrics
+# from . import utils
+# from .base import Network, Model
+#
+# __all__ = ["Network", "Model", "layers", "networks", "activations", "losses", "metrics", "utils"]
