+++ scirex/core/dl/jax_backend/equinox/layers/convolution.py
