+++ scirex/core/dl/jax_backend/equinox/layers/linear.py
