+++ scirex/core/dl/jax_backend/equinox/layers/recurrent.py
