+++ scirex/core/dl/jax_backend/equinox/networks/__init__.py
@@ -1,0 +1,5 @@
+from .mlp import MLP
+from .fcnn import FCNN
+from .gcn import GCN, GCNModel
+
+__all__ = ["MLP", "FCNN", "GCN", "GCNModel"]
