+++ scirex/core/dl/jax_backend/equinox/networks/gcn.py
