+++ scirex/core/dl/jax_backend/equinox/networks/mlp.py
